refactor Promotions
