+++ api/promotions/Promotion.js
@@ -1,62 +1,7 @@
-export default class Promotion {
-  constructor(dataSource, dataTransformer) {
-    this.dataSource = dataSource
-    this.tranformer = dataTransformer
+import AirtableObject from "../lib/AirtableObject"
+
+export default class Promotion extends AirtableObject {
+  constructor(dataSource) {
+    super(dataSource)
   }
-
-  async list() {
-    // returns a promise that resolves to the list of promotions
-    let records = await this.dataSource.list()
-    return this.tranformer.listXform(records)
-  }
-
-  async create(business, text, url) {
-    let record = await this.dataSource.create(business, text, url)
-    return this.tranformer.createXform(record)
-  }
-
-  // read(id) {
-  //   const params = {
-  //     TableName: this.tableName,
-  //     Key: {
-  //       id: id,
-  //     },
-  //   }
-  //   return this.db.get(params).promise()
-  // }
-
-  // update(id, info) {
-  //   // update the data with appropriate identifiers
-  //   const timestamp = new Date().getTime()
-
-  //   var params = {
-  //     TableName: this.tableName,
-  //     Key: {
-  //       id: id,
-  //     },
-  //     UpdateExpression:
-  //       'set competitors = :comp, updatedAt = :updatedAt, #loc = :loc, title = :title',
-  //     ExpressionAttributeValues: {
-  //       ':comp': info.competitors,
-  //       ':loc': info.location,
-  //       ':title': info.title,
-  //       ':updatedAt': timestamp,
-  //     },
-  //     ExpressionAttributeNames: {
-  //       '#loc': 'location',
-  //     },
-  //     ReturnValues: 'ALL_NEW',
-  //   }
-  //   return this.db.update(params).promise()
-  // }
-
-  // delete(id) {
-  //   const params = {
-  //     TableName: this.tableName,
-  //     Key: {
-  //       id: id,
-  //     },
-  //   }
-  //   return this.db.delete(params).promise()
-  // }
 }
